requirements.txt
anyio==4.1.0
asttokens==2.4.1
attrs==23.1.0
binapy==0.7.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
furl==2.1.3
geographiclib==2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
icontract==2.6.6
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
jwskate==0.10.0
MarkupSafe==2.1.3
orderedmultidict==1.0.1
pycparser==2.21
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests_oauth2client==1.3.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
